preparing pvp template - code rework.
