+++ src/main/java/smpvp/smpvp/SMPvp.java
@@ -3,15 +3,16 @@ package smpvp.smpvp;
 import org.bukkit.Bukkit;
 import org.bukkit.event.Listener;
 import org.bukkit.plugin.java.JavaPlugin;
+import smpvp.smpvp.arenas.ArenaManager;
+import smpvp.smpvp.events.GroupFight;
 
 public final class SMPvp extends JavaPlugin implements Listener {
-
 
     @Override
     public void onEnable() {
         // Plugin startup logic
-        Bukkit.getPluginManager().registerEvents(new Events(), this);
-
+        Bukkit.getPluginManager().registerEvents(new GroupFight(), this);
+        ArenaManager.CreateArena("Arena1",2);
 
 
     }
